unsigned int

+++ includes/libft.h
@@ -6,7 +6,7 @@
 /*   By: tmarx <marvin@42.fr>                       +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/10/08 12:09:33 by tmarx             #+#    #+#             */
-/*   Updated: 2019/10/18 11:16:01 by tmarx            ###   ########.fr       */
+/*   Updated: 2019/11/04 09:34:39 by tmarx            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -49,7 +49,7 @@ void			*ft_memmove(void *dst, const void *src, size_t len);
 void			*ft_memchr(const void *s, int c, size_t n);
 int				ft_memcmp(const void *s1, const void *s2, size_t n);
 void			ft_putendl_fd(char const *s, int fd);
-char			*ft_itoa(int nbr);
+char			*ft_itoa(long int nbr);
 char			*ft_substr(char const *s, unsigned int start, size_t len);
 char			*ft_strjoin(char const *s1, char const *s2);
 char			*ft_strtrim(char const *s1, char const *set);
